SharedVIewModel
Added - A brief implementation of the shared view model.
Added - A method to verify that the user input is valid.
Added - A method to parse and lookup a priority based on a given string passed in.
Added - Implementation of assigning a unique color to the spinner textview after a priority is selected.