Input & UI Validation
Added - Implementation of required fields validation for both the title and description.
Added - Implementation of UI validation for both required fields to display a error text message accordingly.
Added - Hide the keyboard once a task is inserted into the database successfully.
Added - A snackbar to be displayed when the required fields are missing.
Added - Adapter creation from resource of strings.
Added - Observer/collect  when navigating back with result with the event system implemented by setting the fragment result with specific keys for request and result before navigating back to the list fragment.