Update Fragment Result Navigation & UI Validation
Added - Displaying the date created of the task being looked at by the user.
Added - Navigated back with result for update and delete actions to present a proper notification message through a snack bar when navigating back to the list fragment of tasks.
Added - User interface validations for when all required fields for updating a task are missing.
Added - Hide keyboard implementation when navigating back to the previous fragment.